[CONTAINER-188] XlsReader で，HSSFSheet#getRow(int) の引数を short にキャストして呼び出しているため，32767 行までしか扱えない．[Seasar-user:13818] https://www.seasar.org/issues/browse/CONTAINER-188 http://ml.seasar.org/archives/seasar-user/2008-April/013818.html
git-svn-id: https://www.seasar.org/svn/s2container/trunk@3566 319488c0-e101-0410-93bc-b5e51f62721a

+++ seasar2/s2-extension/src/main/java/org/seasar/extension/dataset/impl/XlsReader.java
@@ -269,7 +269,7 @@ public class XlsReader implements DataReader, DataSetConstants {
      */
     protected void setupRows(DataTable table, HSSFSheet sheet) {
         for (int i = 1;; ++i) {
-            HSSFRow row = sheet.getRow((short) i);
+            HSSFRow row = sheet.getRow(i);
             if (row == null) {
                 break;
             }
